Move LM engagement functions into utils script
