+++ tests/test_file_schemas.py
@@ -1751,6 +1751,11 @@ class MergeCoverageData:
             StructField(CQCRatings.overall_rating, StringType(), True),
         ]
     )
+
+
+@dataclass
+class LmEngagementUtilsSchemas:
+
     add_columns_for_locality_manager_dashboard_schema = StructType(
         [
             StructField(IndCQC.location_id, StringType(), True),
